// frontend/src/types/Incident.ts
export interface Incident {
    uniqueKey?: string | null;
    createdDate?: Date | null;
    closeDate?: Date | null;
    agency?: string | null;
    agencyName?: string | null;
    complaintType?: string | null;
    descriptor?: string | null;
    locationType?: string | null;
    incidentZip?: string | null;
    incidentAddress?: string | null;
    streetName?: string | null;
    crossStreet1?: string | null;
    crossStreet2?: string | null;
    intersectionStreet1?: string | null;
    intersectionStreet2?: string | null;
    addressType?: string | null;
    city?: string | null;
    landmark?: string | null;
    facilityType?: string | null;
    status?: string | null;
    dueDate?: Date | null;
    resolutionDescription?: string | null;
    resolutionActionUpdatedDate?: Date | null;
    communityBoard?: string | null;
    bbl?: string | null;
    borough?: string | null;
    xCoordinateStatePlane?: number | null;
    yCoordinateStatePlane?: number | null;
    openDataChannelType?: string | null;
    parkFacilityName?: string | null;
    parkBorough?: string | null;
    vehicleType?: string | null;
    taxiCompanyBorough?: string | null;
    taxiPickUpLocation?: string | null;
    bridgeHighwayName?: string | null;
    bridgeHighwayDirection?: string | null;
    roadRamp?: string | null;
    bridgeHighwaySegment?: string | null;
    latitude?: number | null;
    longitude?: number | null;
    location?: string | null;
}